+++ zerver/tests/test_home.py
@@ -107,6 +107,7 @@ class HomeTest(ZulipTestCase):
         "max_message_id",
         "message_content_in_email_notifications",
         "muted_topics",
+        "muted_users",
         "narrow",
         "narrow_stream",
         "needs_tutorial",
@@ -261,7 +262,7 @@ class HomeTest(ZulipTestCase):
             set(result["Cache-Control"].split(", ")), {"must-revalidate", "no-store", "no-cache"}
         )
 
-        self.assert_length(queries, 39)
+        self.assert_length(queries, 40)
         self.assert_length(cache_mock.call_args_list, 5)
 
         html = result.content.decode("utf-8")
@@ -341,7 +342,7 @@ class HomeTest(ZulipTestCase):
                 result = self._get_home_page()
                 self.check_rendered_logged_in_app(result)
                 self.assert_length(cache_mock.call_args_list, 6)
-            self.assert_length(queries, 36)
+            self.assert_length(queries, 37)
 
     def test_num_queries_with_streams(self) -> None:
         main_user = self.example_user("hamlet")
@@ -372,7 +373,7 @@ class HomeTest(ZulipTestCase):
         with queries_captured() as queries2:
             result = self._get_home_page()
 
-        self.assert_length(queries2, 34)
+        self.assert_length(queries2, 35)
 
         # Do a sanity check that our new streams were in the payload.
         html = result.content.decode("utf-8")